+++ sdk/ai/Azure.AI.Projects/tests/AgentsTests.cs
@@ -8,6 +8,7 @@ using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
@@ -17,6 +18,7 @@ using Azure.AI.Extensions.OpenAI;
 using Azure.AI.Projects.Agents;
 using Azure.AI.Projects.Tests.Utils;
 using Microsoft.ClientModel.TestFramework;
+using Microsoft.Extensions.AI;
 using NUnit.Framework;
 using OpenAI;
 using OpenAI.Containers;
@@ -689,6 +691,7 @@ public class AgentsTests : AgentsTestBase
     [TestCase(ToolType.FileSearch)]
     [TestCase(ToolType.ImageGeneration)]
     [TestCase(ToolType.WebSearch)]
+    [TestCase(ToolType.WebSearchCustom)]
     [TestCase(ToolType.Memory)]
     [TestCase(ToolType.AzureAISearch)]
     [TestCase(ToolType.BingGrounding)]
@@ -759,7 +762,7 @@ public class AgentsTests : AgentsTestBase
                 Assert.That(Regex.Match(response.GetOutputText().ToLower(), expectedResponse.ToLower()).Success, Is.True, $"The output: \"{response.GetOutputText()}\" does not contain {expectedResponse}");
             }
         }
-        if (toolType == ToolType.AzureAISearch | toolType == ToolType.BingGrounding | toolType == ToolType.BingGroundingCustom | toolType == ToolType.Sharepoint | toolType == ToolType.MicrosoftFabric)
+        if (toolType == ToolType.AzureAISearch | toolType == ToolType.BingGrounding | toolType == ToolType.BingGroundingCustom | toolType == ToolType.Sharepoint | toolType == ToolType.MicrosoftFabric | toolType == ToolType.WebSearch | toolType == ToolType.WebSearchCustom)
         {
             bool isUriCitationFound = false;
 
@@ -787,6 +790,8 @@ public class AgentsTests : AgentsTestBase
     [TestCase(ToolType.CodeInterpreter)]
     [TestCase(ToolType.CodeInterpreterGen)]
     [TestCase(ToolType.Memory)]
+    [TestCase(ToolType.WebSearch)]
+    [TestCase(ToolType.WebSearchCustom)]
     [TestCase(ToolType.AzureAISearch)]
     [TestCase(ToolType.BingGrounding)]
     [TestCase(ToolType.BingGroundingCustom)]
@@ -849,13 +854,13 @@ public class AgentsTests : AgentsTestBase
                             }
                         }
                     }
-                    if (toolType == ToolType.AzureAISearch | toolType == ToolType.BingGrounding | toolType == ToolType.BingGroundingCustom | toolType == ToolType.Sharepoint | toolType == ToolType.MicrosoftFabric)
+                    if (toolType == ToolType.AzureAISearch | toolType == ToolType.BingGrounding | toolType == ToolType.BingGroundingCustom | toolType == ToolType.Sharepoint | toolType == ToolType.MicrosoftFabric | toolType == ToolType.WebSearch | toolType == ToolType.WebSearchCustom)
                     {
-                        annotationMet = ContainsAnnotation(itemDoneUpdate.Item, toolType);
+                        annotationMet |= ContainsAnnotation(itemDoneUpdate.Item, toolType);
                     }
                     if (toolType == ToolType.CodeInterpreter)
                     {
-                        annotationMet = await ContainsDownloadableFileAnnotation(itemDoneUpdate.Item, projectClient);
+                        annotationMet |= await ContainsDownloadableFileAnnotation(itemDoneUpdate.Item, projectClient);
                     }
                 }
                 else
@@ -1536,6 +1541,7 @@ public class AgentsTests : AgentsTestBase
 
     private bool ContainsAnnotation(ResponseItem item, ToolType type)
     {
+        StringBuilder sbAnnotations = new();
         bool isUriCitationFound = false;
         if (item is MessageResponseItem messageItem)
         {
@@ -1546,16 +1552,40 @@ public class AgentsTests : AgentsTestBase
                     if (annotation is UriCitationMessageAnnotation uriAnnotation)
                     {
                         isUriCitationFound = true;
-                        Assert.That(uriAnnotation.Title, Does.Contain(ExpectedAnnotationTitle[type]), $"Wrong citation title {uriAnnotation.Title}, should be \"product_info_7.md\"");
-                        // The next check is disabled, because of an ADO issue 4836442.
-                        // Assert.That(uriAnnotation.Uri, Does.Contain("www.microsoft.com"), $"Wrong citation title {uriAnnotation.Uri}, should be \"www.microsoft.com\"");
+                        if (ExpectedAnnotationTitle.TryGetValue(type,out string expectedTitle))
+                        {
+                            if (uriAnnotation.Title.ToLower().Contains(expectedTitle.ToLower()))
+                            {
+                                isUriCitationFound = true;
+                                break;
+                            }
+                            else
+                            {
+                                sbAnnotations.Append($"[{uriAnnotation.Title}]({uriAnnotation.Uri})\n");
+                                // The next check is disabled, because of an ADO issue 4836442.
+                                // Assert.That(uriAnnotation.Uri, Does.Contain("www.microsoft.com"), $"Wrong citation title {uriAnnotation.Uri}, should be \"www.microsoft.com\"");
+                            }
+                        }
+                        else
+                        {
+                            isUriCitationFound = true;
+                            break;
+                        }
                     }
                     else
                     {
                         Assert.Fail($"Found unexpected annotation {annotation}");
                     }
                 }
+                if (isUriCitationFound)
+                {
+                    break;
+                }
             }
+        }
+        if (!isUriCitationFound && sbAnnotations.Length > 0)
+        {
+            Assert.Fail($"Found wrong citations:\n{sbAnnotations}");
         }
         return isUriCitationFound;
     }