+++ sdk/ai/Azure.AI.Projects/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Breaking Changes
 * The `Azure.AI.Projects.OpenAI` package was renamed to `Azure.AI.Extensions.OpenAI`
-* The Agents administration was moved to the `Azure.AI.Projects.Agent` package.
+* The Agents administration was moved to the `Azure.AI.Projects.Agents` package.
 * OpenAI dependency was upgraded to 2.9.1.
 
 ## 2.0.0-beta.1 (2026-02-24)